# Null Pointer Optimization in Node children