## This PR

Reworks `Node` children from `Rc<Tree>` (which is roughly `Rc<Option<Node>>`) to `Option<Rc<Node>>` to take advantage of the null pointer optimization there.

This is done in two commits. The first one introduces a new `Child` struct to make it easier to review the second commit which adds the `Option`al structure.

It's quite possible the `Child` struct can be removed at this point. Benchmarks for that should be investigated.

<details>
<summary>Benchmarks</summary>

Benchmarks for this branch (both commits together), with outliers removed, are:

```
find/unbalanced/6       time:   [3.2420 ns 3.2549 ns 3.2667 ns]
                        change: [+2.4473% +3.8006% +4.7412%] (p = 0.00 < 0.05)
                        Performance has regressed.
find/balanced/6         time:   [3.2638 ns 3.2666 ns 3.2693 ns]
                        change: [+4.0062% +4.4559% +4.8542%] (p = 0.00 < 0.05)
                        Performance has regressed.
find/unbalanced/126     time:   [7.0310 ns 7.0371 ns 7.0450 ns]
                        change: [-21.931% -21.745% -21.522%] (p = 0.00 < 0.05)
                        Performance has improved.
find/balanced/126       time:   [8.3811 ns 8.3910 ns 8.4046 ns]
                        change: [-22.623% -21.897% -21.012%] (p = 0.00 < 0.05)
                        Performance has improved.
find/unbalanced/2046    time:   [13.274 ns 13.276 ns 13.278 ns]
                        change: [-15.056% -14.603% -13.979%] (p = 0.00 < 0.05)
                        Performance has improved.
find/balanced/2046      time:   [14.408 ns 14.423 ns 14.439 ns]
                        change: [-19.763% -19.218% -18.512%] (p = 0.00 < 0.05)
                        Performance has improved.
find/unbalanced/32766   time:   [24.049 ns 24.097 ns 24.167 ns]
                        change: [-5.7962% -5.2063% -4.4081%] (p = 0.00 < 0.05)
                        Performance has improved.
find/balanced/32766     time:   [21.755 ns 21.761 ns 21.768 ns]
                        change: [-19.085% -18.789% -18.541%] (p = 0.00 < 0.05)
                        Performance has improved.

delete/unbalanced/6     time:   [51.879 ns 51.904 ns 51.932 ns]
                        change: [-26.296% -25.769% -25.140%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/balanced/6       time:   [52.701 ns 52.770 ns 52.846 ns]
                        change: [-25.499% -24.855% -24.279%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/unbalanced/126   time:   [180.46 ns 180.63 ns 180.83 ns]
                        change: [-8.7991% -8.2582% -7.7966%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/balanced/126     time:   [211.67 ns 212.47 ns 214.01 ns]
                        change: [-7.7736% -7.1557% -6.4458%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/unbalanced/2046  time:   [360.06 ns 360.20 ns 360.36 ns]
                        change: [-10.101% -9.2309% -8.0783%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/balanced/2046    time:   [500.53 ns 501.32 ns 502.46 ns]
                        change: [-4.3721% -3.1968% -1.9218%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/unbalanced/32766 time:   [594.57 ns 596.50 ns 598.92 ns]
                        change: [-7.2811% -6.0005% -4.8648%] (p = 0.00 < 0.05)
                        Performance has improved.
delete/balanced/32766   time:   [634.35 ns 635.54 ns 637.18 ns]
                        change: [-2.9938% -2.1631% -1.3887%] (p = 0.00 < 0.05)
                        Performance has improved.

insert/unbalanced/6     time:   [131.64 ns 132.16 ns 132.87 ns]
                        change: [-22.099% -18.188% -12.788%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/balanced/6       time:   [132.10 ns 136.08 ns 141.57 ns]
                        change: [-37.461% -33.748% -29.746%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/unbalanced/126   time:   [259.74 ns 260.07 ns 260.40 ns]
                        change: [-27.419% -26.643% -25.980%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/balanced/126     time:   [472.11 ns 473.84 ns 476.33 ns]
                        change: [-8.9275% -8.1607% -7.3933%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/unbalanced/2046  time:   [475.68 ns 476.29 ns 476.95 ns]
                        change: [-16.758% -15.418% -14.318%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/balanced/2046    time:   [599.24 ns 599.93 ns 600.68 ns]
                        change: [-7.7128% -5.5058% -1.9537%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/unbalanced/32766 time:   [691.80 ns 692.33 ns 692.94 ns]
                        change: [-15.379% -13.663% -11.858%] (p = 0.00 < 0.05)
                        Performance has improved.
insert/balanced/32766   time:   [895.05 ns 895.88 ns 896.99 ns]
                        change: [-7.9344% -7.1838% -6.3918%] (p = 0.00 < 0.05)
                        Performance has improved.

find-miss/unbalanced/6  time:   [2.7059 ns 2.7247 ns 2.7556 ns]
                        change: [-28.090% -27.424% -26.830%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/balanced/6    time:   [2.6940 ns 2.7014 ns 2.7097 ns]
                        change: [-28.395% -27.776% -27.273%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/unbalanced/126
                        time:   [7.7050 ns 7.7136 ns 7.7259 ns]
                        change: [-18.283% -17.956% -17.620%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/balanced/126  time:   [8.5930 ns 8.6019 ns 8.6135 ns]
                        change: [-22.010% -21.761% -21.500%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/unbalanced/2046
                        time:   [12.782 ns 12.791 ns 12.801 ns]
                        change: [-20.254% -19.833% -19.235%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/balanced/2046 time:   [14.944 ns 14.952 ns 14.964 ns]
                        change: [-17.540% -17.080% -16.777%] (p = 0.00 < 0.05)
                        Performance has improved.
find-miss/unbalanced/32766
                        time:   [26.878 ns 26.893 ns 26.909 ns]
                        change: [+5.2232% +5.7561% +6.1261%] (p = 0.00 < 0.05)
                        Performance has regressed.
find-miss/balanced/32766
                        time:   [20.402 ns 20.411 ns 20.424 ns]
                        change: [-25.318% -24.863% -24.356%] (p = 0.00 < 0.05)
                        Performance has improved.

delete-miss/unbalanced/6
                        time:   [82.394 ns 82.628 ns 82.946 ns]
                        change: [-18.087% -17.520% -17.060%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/balanced/6  time:   [82.500 ns 82.728 ns 83.007 ns]
                        change: [-18.584% -17.829% -17.125%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/unbalanced/126
                        time:   [207.83 ns 208.21 ns 208.79 ns]
                        change: [-10.267% -9.4155% -8.5655%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/balanced/126
                        time:   [238.73 ns 239.41 ns 240.38 ns]
                        change: [-43.644% -43.171% -42.496%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/unbalanced/2046
                        time:   [406.43 ns 406.77 ns 407.15 ns]
                        change: [-9.6620% -8.6836% -7.9487%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/balanced/2046
                        time:   [528.08 ns 529.01 ns 530.28 ns]
                        change: [-5.1771% -3.9000% -3.0746%] (p = 0.00 < 0.05)
                        Performance has improved.
delete-miss/unbalanced/32766
                        time:   [647.03 ns 650.77 ns 656.06 ns]
                        change: [-3.0026% -1.3391% +0.9399%] (p = 0.14 > 0.05)
                        No change in performance detected.
delete-miss/balanced/32766
                        time:   [678.55 ns 678.85 ns 679.21 ns]
                        change: [-19.548% -18.782% -17.942%] (p = 0.00 < 0.05)
                        Performance has improved.
```

It looks like the regressions are spurious since improvements can be found
in the same class of benchmark with greater nodes.

</details>